update lib/storage/base to latest implementation
